refactor: BasePopUpView 구현
- BasePopUpView 구현
- 상속하여 사용하도록 변경
- dragHandleBar 이미지에서 UIView 사용으로 변경

#87
