+++ Bridge/Sources/Presentation/Tab/Main/CreateProject/MemberDetailInput/AddTechTag/AddTechTagPopUpView.swift
@@ -12,36 +12,8 @@ import RxSwift
 import RxCocoa
 
 /// 기술스택을 추가하는 뷰
-final class AddTechTagPopUpView: BaseView {
+final class AddTechTagPopUpView: BasePopUpView {
     // MARK: - UI
-    private let rootFlexContainer: UIView = {
-        let view = UIView()
-        view.backgroundColor = BridgeColor.gray10
-        view.clipsToBounds = true
-        view.layer.cornerRadius = 12
-        view.layer.maskedCorners = [.layerMinXMinYCorner, .layerMaxXMinYCorner]
-        
-        return view
-    }()
-    
-    private let dragHandleImageView: UIImageView = {
-        let imageView = UIImageView()
-        imageView.image = UIImage(named: "draghandle")
-        imageView.backgroundColor = .clear
-        imageView.contentMode = .scaleAspectFit
-        
-        return imageView
-    }()
-    
-    private let stackLabel: UILabel = {
-        let label = UILabel()
-        label.text = "스택"
-        label.font = BridgeFont.subtitle1.font
-        label.textColor = BridgeColor.gray1
-        
-        return label
-    }()
-    
     private lazy var collectionView: UICollectionView = {
         let collectionView = UICollectionView(frame: .zero, collectionViewLayout: configureLayout())
         collectionView.backgroundColor = .clear
@@ -51,18 +23,10 @@ final class AddTechTagPopUpView: BaseView {
         return collectionView
     }()
     
-    private let completeButton: BridgeButton = {
-        let button = BridgeButton(
-            title: "완료",
-            font: BridgeFont.button1.font,
-            backgroundColor: BridgeColor.gray4
-        )
-        button.isEnabled = false
-        
-        return button
-    }()
-    
     // MARK: - Properties
+    override var containerHeight: CGFloat { 576 }
+    override var dismissYPosition: CGFloat { 300 }
+    
     /// 선택된 분야를 전달받으면
     /// willSet - 데이터소스를 결정 할 타입을 저장
     /// didSet - 데이터소스를 바인딩하여 CollectionView Cell을 구현(CollectionViewDataSource 역할)
@@ -118,12 +82,12 @@ final class AddTechTagPopUpView: BaseView {
     
     // MARK: - Layout
     override func configureLayouts() {
-        addSubview(rootFlexContainer)
+        super.configureLayouts()
         
         rootFlexContainer.flex.define { flex in
-            flex.addItem(dragHandleImageView).alignSelf(.center).width(25).height(7).marginTop(10)
+            flex.addItem(dragHandleBar).alignSelf(.center).width(27).height(5).marginTop(10)
             
-            flex.addItem(stackLabel).width(32).height(22).marginTop(30).marginLeft(16)
+            flex.addItem(titleLabel).width(32).height(22).marginTop(30).marginLeft(16)
             
             flex.addItem().backgroundColor(BridgeColor.gray8).height(1).marginTop(7)
             
@@ -136,16 +100,12 @@ final class AddTechTagPopUpView: BaseView {
     
     override func layoutSubviews() {
         super.layoutSubviews()
-        rootFlexContainer.pin.below(of: self).width(100%).height(576)
-        rootFlexContainer.flex.layout()
     }
     
     // MARK: - configure
     override func configureAttributes() {
-        backgroundColor = UIColor(red: 0, green: 0, blue: 0, alpha: 0.3)
-        
-        let panGesture = UIPanGestureRecognizer(target: self, action: #selector(handlePanGesture))
-        rootFlexContainer.addGestureRecognizer(panGesture)
+        super.configureAttributes()
+        titleLabel.text = "스택"
     }
 }
 
@@ -170,74 +130,6 @@ extension AddTechTagPopUpView {
         section.interGroupSpacing = 14
         
         return UICollectionViewCompositionalLayout(section: section)
-    }
-}
-
-
-// MARK: - PanGesture
-extension AddTechTagPopUpView {
-    @objc
-    private func handlePanGesture(sender: UIPanGestureRecognizer) {
-        let translation = sender.translation(in: rootFlexContainer)
-        let velocity = sender.velocity(in: rootFlexContainer)
-        
-        let currentTranslationY: CGFloat = -576  // 초기 컨테이너의 오프셋
-        let calculatedTranslationY = currentTranslationY + translation.y  // 초기 오프셋을 고려한 결과값
-        
-        switch sender.state {
-        case .changed:
-            // 아래로 드래그하거나 최대 높이 이내에서만 transform을 적용
-            if calculatedTranslationY > currentTranslationY && calculatedTranslationY <= 0 {
-                rootFlexContainer.transform = CGAffineTransform(translationX: 0, y: calculatedTranslationY)
-            }
-            
-        case .ended:
-            if velocity.y > 1500 || calculatedTranslationY > -300 {
-                hide()
-                
-            } else {  // 원상복구
-                UIView.animate(withDuration: 0.2) {
-                    self.rootFlexContainer.transform = CGAffineTransform(translationX: 0, y: currentTranslationY)
-                }
-            }
-            
-        default:
-            break
-        }
-    }
-}
-
-// MARK: - Show & Hide
-extension AddTechTagPopUpView {
-    func show() {
-        setLayout()
-        
-        UIView.animate(withDuration: 0.2, animations: { [weak self] in
-            guard let self else { return }
-            self.rootFlexContainer.transform = CGAffineTransform(translationX: 0, y: -576)
-        })
-    }
-
-    private func setLayout() {
-        let window = UIWindow.visibleWindow() ?? UIWindow()
-        window.addSubview(self)
-        window.bringSubviewToFront(self)
-        
-        pin.all()
-    }
-    
-    private func hide() {
-        UIView.animate(
-            withDuration: 0.2,
-            animations: { [weak self] in
-                guard let self else { return }
-                self.rootFlexContainer.transform = .identity
-                
-            }, completion: { [weak self] _ in
-                guard let self else { return }
-                self.removeFromSuperview()
-            }
-        )
     }
 }
